Refactor 2025-08-21

+++ scripts/remove/01_nginx.sh
@@ -1,10 +1,15 @@
 #!/bin/bash
 
-set -e
+set -euo pipefail
 # set -x
 
 . ./.env
 . ./scripts/functions.sh
+
+if [[ $(kubectl get namespace | grep "${INGRESS_NGINX_NAMESPACE}" | wc -l) -lt 1 ]]; then
+    echo "Namespace ${INGRESS_NGINX_NAMESPACE} does not exist, skipping NGINX deletion"
+    exit 0
+fi
 
 helm uninstall ingress-nginx \
     --namespace ${INGRESS_NGINX_NAMESPACE}
