+++ scripts/remove/02_vault.sh
@@ -1,17 +1,22 @@
 #!/bin/bash
 
-set -e
+set -euo pipefail
 # set -x
 
 . ./.env
 . ./scripts/functions.sh
 
-kubectl -n ${VAULT_NAMESPACE} delete \
+if [[ $(kubectl get namespace | grep "${VAULT_NAMESPACE}" | wc -l) -lt 1 ]]; then
+    echo "Namespace ${VAULT_NAMESPACE} does not exist, skipping Vault deletion"
+    exit 0
+fi
+
+kubectl -n "${VAULT_NAMESPACE}" delete \
     --ignore-not-found=true \
     Ingress/vault
 
 helm uninstall vault \
-    --namespace ${VAULT_NAMESPACE}
+    --namespace "${VAULT_NAMESPACE}"
 
-kubectl delete namespace ${VAULT_NAMESPACE} \
+kubectl delete namespace "${VAULT_NAMESPACE}" \
     --ignore-not-found=true
