+++ scripts/remove/10_cp_certs.sh
@@ -1,10 +1,15 @@
 #!/bin/bash
 
-set -e
+set -euo pipefail
 # set -x
 
 . ./.env
 . ./scripts/functions.sh
+
+if [[ $(kubectl get namespace | grep "${NAMESPACE}" | wc -l) -lt 1 ]]; then
+    echo "Namespace ${NAMESPACE} does not exist, skipping CP certs deletion"
+    exit 0
+fi
 
 # Basic infra
 kubectl -n "${NAMESPACE}" delete \
