+++ scripts/remove/50_pipeline_demo.sh
@@ -1,10 +1,21 @@
 #!/bin/bash
 
-set -e
+set -euo pipefail
 # set -x
 
 . ./.env
 . ./scripts/functions.sh
+
+if [[ $(kubectl get namespace | grep "${NAMESPACE}" | wc -l) -lt 1 ]]; then
+    echo "Namespace ${NAMESPACE} does not exist, skipping pipeline demo deletion"
+    exit 0
+fi
+
+if [[ $(kubectl -n "${NAMESPACE}" get pod -l app=connect | grep "1/1" | wc -l) -lt 1 ]]; then
+    echo "Connect is not running, removing finalizers to skip validation"
+    remove_finalizer "${NAMESPACE}" connector/elasticsearch-sink
+    remove_finalizer "${NAMESPACE}" connector/wikipedia-sse-source
+fi
 
 kubectl -n "${NAMESPACE}" delete \
     --ignore-not-found=true \
